fix: replace `merge()` with `Object.assign()`
## Background

In my tests - I often spin up servers as part of `beforeEach`. In the particular area where I encountered problems is that my test server kept track of incoming requests by a mutable array. Couldn't for the life of me understand why the array was constantly empty, but I think this was the culprit.

To be even more specific, I was testing a websocket trigger/receiver - so I in my tests I dynamically use `http.createServer()` & later I want to assert that the server received events as after doing specific actions in my app.